Remove old event
